# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DoReMi - Plateforme scolaire numérique",
        "description": "Bienvenue dans la documentation de l'API **DoReMi**, une plateforme complète de gestion scolaire hybride.\\n\\n\n        ## Modules couverts :\\n\n        - Gestion des écoles et utilisateurs (élèves, enseignants, parents)\\n\n        - Cours en ligne et en présentiel\\n\n        - Suivi des notes et bulletins\\n\n        - Gestion des documents et devoirs\\n\n        - Actualités, emploi du temps, messagerie\\n\n        - Gestion des stages : demandes & offres\\n\\n\n        Cette documentation est à jour pour l'environnement de développement.",
        "termsOfService": "https://doremi-app.test/conditions",
        "contact": {
            "name": "Support Technique DoReMi",
            "email": "support@doremi-app.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/subscriptions/{userId}": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Obtenir les abonnements d'un utilisateur",
                "operationId": "8a65e9bfdd68a8c8d189d5a2b0d4cdd3",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abonnements récupérés avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/subscriptions": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Créer un nouvel abonnement",
                "operationId": "55785339e6eafa3b4c917ada608130de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "plan_type": {
                                        "type": "string",
                                        "enum": [
                                            "basic",
                                            "premium",
                                            "enterprise"
                                        ]
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "default": "EUR"
                                    },
                                    "duration_months": {
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "auto_renew": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Abonnement créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/subscriptions/{id}": {
            "put": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Mettre à jour un abonnement",
                "operationId": "177bb0f71099e1b87a2e5d8f15546129",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_type": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "auto_renew": {
                                        "type": "boolean"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Abonnement mis à jour avec succès"
                    },
                    "404": {
                        "description": "Abonnement non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Supprimer un abonnement",
                "operationId": "1c3de098cff6f883a313b910c1f9d621",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abonnement supprimé avec succès"
                    },
                    "404": {
                        "description": "Abonnement non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/payments": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Traiter un paiement",
                "operationId": "6ed7838e6c677c90f52dd3ba24f13269",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "subscription_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "currency": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "transaction_id": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "failed",
                                            "refunded"
                                        ]
                                    },
                                    "metadata": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement traité avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/history/{userId}": {
            "get": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Obtenir l'historique des paiements d'un utilisateur",
                "operationId": "2cc267f05dc8d276bf7bbf10cf11effc",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historique des paiements récupéré avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/subscriptions/{id}/cancel": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Annuler un abonnement",
                "operationId": "bf05987435802006ee25933d296d3709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abonnement annulé avec succès"
                    },
                    "404": {
                        "description": "Abonnement non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/subscriptions/{id}/renew": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Renouveler un abonnement",
                "operationId": "f6163957a3ca3bdd0fb0331c9c32aee7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abonnement renouvelé avec succès"
                    },
                    "404": {
                        "description": "Abonnement non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}